requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
importlib-metadata==7.0.0
Markdown==3.5.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
zipp==3.17.0
